feat: klondike from v0.dev
I literally just get v0.dev to iterate with fairly basic prompts, and fixed one linting error about strict typing any[] to Card[]

Prompt 1
=========
I would like a klondike clone, with two game modes. 3 card and 1 card pile. I would like keyboard ui instead of thinking about drag and drop, 7 columns for placing cards, starting the game with n 1 cards initially from left-most column to the right-most column, which has 7 cards. Cards can be in one of two states, face-up or face-down

Prompt 2
=========
while clever it does not tell me how to move a drawn card to the tableau

Prompt 3
=========
Should I be able to move waste card to the tableau when empty? If not currently, lets account for this

Prompt 4
=========
The last edit did not work. Also F is only working from waste, so if I get an ace in tableau, then I can't send to foundation.

Prompt 5
=========
pressing N should start a new game

Prompt 6
=========
okay after foundation is correctly moved to, I feel like there should be no selected card

Prompt 7
=========
it looks like the game doesn't know when someone wins. A Congratulatory message spanning from 1st tableau column to 7th which is roughly 5 cards in depth and says "Congratulations! You Won" would be cool. It's font weight should be 400, sans serif

Prompt 8
=========
the css grid setup is causing the border to be larger than it needs to be for foundation cards when selected. I believe we overcome this by drawing the border around an item within an item, so that it is not grid relative. This seems to work for the cards themselves as they display with the correct border